--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/DrivePowerConstants.java
@@ -1,9 +1,0 @@
-package org.firstinspires.ftc.teamcode.opmode;
-
-import com.acmerobotics.dashboard.config.Config;
-
-@Config
-public class DrivePowerConstants {
-    public static double lowDrivePower = 0.3;
-    public static double highDrivePower = 0.6;
-}
