+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/teleop/GamepadButton.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmode;
+package org.firstinspires.ftc.teamcode.opmode.teleop;
 
 public enum GamepadButton {
     A,
