+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/teleop/GamepadControllerBase.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmode;
+package org.firstinspires.ftc.teamcode.opmode.teleop;
 
 import com.qualcomm.robotcore.hardware.Gamepad;
